\name{GSS.2016}
\alias{GSS.2016}
\docType{data}
\title{
Subset of data from the General Social Survey from 2016. Data were analyzed in Schoon, Melamed, and Breiger (2024).
}
\description{
Subset of data from the General Social Survey from 2016. Data were analyzed in Schoon, Melamed, and Breiger (2024). Full details on the variable selection and source information is available therein.}
\usage{data("GSS.2016")}
\format{
  A data frame with 2867 observations on the following 27 variables.
  \describe{
    \item{\code{sclass}}{a numeric vector}
    \item{\code{fulltime}}{a numeric vector}
    \item{\code{retired}}{a numeric vector}
    \item{\code{hrsworked}}{a numeric vector}
    \item{\code{occprestige}}{a numeric vector}
    \item{\code{occprestige_partner}}{a numeric vector}
    \item{\code{occprestige_mother}}{a numeric vector}
    \item{\code{occprestige_father}}{a numeric vector}
    \item{\code{children}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{educ}}{a numeric vector}
    \item{\code{paeduc}}{a numeric vector}
    \item{\code{maeduc}}{a numeric vector}
    \item{\code{speduc}}{a numeric vector}
    \item{\code{babs}}{a numeric vector}
    \item{\code{female}}{a numeric vector}
    \item{\code{white}}{a numeric vector}
    \item{\code{black}}{a numeric vector}
    \item{\code{other}}{a numeric vector}
    \item{\code{income}}{a numeric vector}
    \item{\code{republican}}{a numeric vector}
    \item{\code{conservative}}{a numeric vector}
    \item{\code{environment}}{a numeric vector}
    \item{\code{helpblackpeople}}{a numeric vector}
    \item{\code{science}}{a numeric vector}
    \item{\code{govequalwealth}}{a numeric vector}
    \item{\code{pclass}}{a numeric vector}
  }
}
\references{
Schoon, Eric W., David Melamed, and Ronald L. Breiger. 2024. Regression Inside Out. NY: Cambridge University Press.
}
\examples{
data(GSS.2016)
head(GSS.2016)
}
\keyword{datasets}
