\name{SchneiderAndMakszin06}
\alias{SchneiderAndMakszin06}
\docType{data}
\title{
Subset of replication data from Schneider and Makszin (2014).
}
\description{
Subset of replication data from Schneider and Makszin (2014). Data were analyzed in Schoon, Melamed, and Breiger (2024). Full details on the variable selection and source information is available therein.
}
\usage{data("SchneiderAndMakszin06")}
\format{
  A data frame with 30 observations on the following 36 variables.
  \describe{
    \item{\code{id}}{a character vector}
    \item{\code{country}}{a character vector}
    \item{\code{year}}{a numeric vector}
    \item{\code{fde}}{a numeric vector}
    \item{\code{fde_cilb}}{a numeric vector}
    \item{\code{fde_ciub}}{a numeric vector}
    \item{\code{wcoord}}{a numeric vector}
    \item{\code{govint}}{a numeric vector}
    \item{\code{ud}}{a numeric vector}
    \item{\code{epl}}{a numeric vector}
    \item{\code{socexp}}{a numeric vector}
    \item{\code{eduexp}}{a numeric vector}
    \item{\code{vet_un}}{a numeric vector}
    \item{\code{lmexp}}{a numeric vector}
    \item{\code{wagecov}}{a numeric vector}
    \item{\code{vet_isced3}}{a numeric vector}
    \item{\code{eduexp_pri}}{a numeric vector}
    \item{\code{edu_terenr}}{a numeric vector}
    \item{\code{vt_reg}}{a numeric vector}
    \item{\code{vt_vap}}{a numeric vector}
    \item{\code{compvote}}{a numeric vector}
    \item{\code{fde2}}{a numeric vector}
    \item{\code{low_fde_l}}{a numeric vector}
    \item{\code{high_fde_l}}{a numeric vector}
    \item{\code{high_wc_l}}{a numeric vector}
    \item{\code{high_int_l}}{a numeric vector}
    \item{\code{high_ud_l}}{a numeric vector}
    \item{\code{high_epl_l}}{a numeric vector}
    \item{\code{high_socx_l}}{a numeric vector}
    \item{\code{high_edux_l}}{a numeric vector}
    \item{\code{high_lmx_l}}{a numeric vector}
    \item{\code{high_vet_l}}{a numeric vector}
    \item{\code{p1_y}}{a numeric vector}
    \item{\code{p2_y}}{a numeric vector}
    \item{\code{p3_y}}{a numeric vector}
    \item{\code{sol_y}}{a numeric vector}
  }
}
\references{
Schneider, Carsten Q., and Kristin Makszin. 2014. "Forms of welfare capitalism and education-based participatory inequality." Socio-Economic Review 12(2): 437-462.
Schoon, Eric W., David Melamed, and Ronald L. Breiger. 2024. Regression Inside Out. NY: Cambridge University Press.
}
\examples{
data(SchneiderAndMakszin06)
head(SchneiderAndMakszin06)
}
\keyword{datasets}
