\name{Wimmer_et_al_EPR}
\alias{Wimmer_et_al_EPR}
\docType{data}
\title{
Subset of replication data from Wimmer, Cederman, and Min (2009).
}
\description{
Subset of replication data from Wimmer, Cederman, and Min (2009). Data were analyzed in Schoon, Melamed, and Breiger (2024). Full details on the variable selection and source information is available therein.}
\usage{data("Wimmer_et_al_EPR")}
\format{
  A data frame with 7908 observations on the following 80 variables.
  \describe{
    \item{\code{yearc}}{a numeric}
    \item{\code{year}}{a numeric}
    \item{\code{cowcode}}{a numeric}
    \item{\code{country}}{a character}
    \item{\code{gdpcap}}{a numeric}
    \item{\code{gdpcapl}}{a numeric}
    \item{\code{oilpc}}{a numeric}
    \item{\code{oilpcl}}{a numeric}
    \item{\code{popavg}}{a numeric}
    \item{\code{lpopl}}{a numeric}
    \item{\code{ethfrac}}{a numeric}
    \item{\code{western}}{a numeric}
    \item{\code{eeurop}}{a numeric}
    \item{\code{lamerica}}{a numeric}
    \item{\code{ssafrica}}{a numeric}
    \item{\code{asia}}{a numeric}
    \item{\code{nafrme}}{a numeric}
    \item{\code{lmtnest}}{a numeric}
    \item{\code{polity2}}{a numeric}
    \item{\code{polity}}{a numeric}
    \item{\code{anoc}}{a numeric}
    \item{\code{anocl}}{a numeric}
    \item{\code{democ}}{a numeric}
    \item{\code{democl}}{a numeric}
    \item{\code{regchg3}}{a numeric}
    \item{\code{pimppast}}{a numeric}
    \item{\code{groups}}{a numeric}
    \item{\code{egipgrps}}{a numeric}
    \item{\code{exclgrps}}{a numeric}
    \item{\code{exclpop}}{a numeric}
    \item{\code{lrexclpop}}{a numeric}
    \item{\code{ttlpop}}{a numeric}
    \item{\code{discpop}}{a numeric}
    \item{\code{pwrlpop}}{a numeric}
    \item{\code{olppop}}{a numeric}
    \item{\code{olpspop}}{a numeric}
    \item{\code{jppop}}{a numeric}
    \item{\code{sppop}}{a numeric}
    \item{\code{dompop}}{a numeric}
    \item{\code{monpop}}{a numeric}
    \item{\code{maxexclpop}}{a numeric}
    \item{\code{maxegippop}}{a numeric}
    \item{\code{maxpop}}{a numeric}
    \item{\code{newonset}}{a numeric}
    \item{\code{newethonset}}{a numeric}
    \item{\code{newhionset}}{a numeric}
    \item{\code{newethhionset}}{a numeric}
    \item{\code{onsetstatus}}{a numeric}
    \item{\code{onsetstatus2}}{a numeric}
    \item{\code{actoraim}}{a numeric}
    \item{\code{actoraim2}}{a numeric}
    \item{\code{ongoingwarl}}{a numeric}
    \item{\code{ongoinghiwarl}}{a numeric}
    \item{\code{newonset2}}{a numeric}
    \item{\code{newhionset2}}{a numeric}
    \item{\code{newethonset2}}{a numeric}
    \item{\code{warlfl}}{a numeric}
    \item{\code{onsetfl}}{a numeric}
    \item{\code{ethonsetfl}}{a numeric}
    \item{\code{onsetfl2}}{a numeric}
    \item{\code{ethonsetfl2}}{a numeric}
    \item{\code{warstns2}}{a numeric}
    \item{\code{warstns1}}{a numeric}
    \item{\code{atwarnsl}}{a numeric}
    \item{\code{npeaceyears}}{a numeric}
    \item{\code{nspline1}}{a numeric}
    \item{\code{nspline2}}{a numeric}
    \item{\code{nspline3}}{a numeric}
    \item{\code{hpeaceyears}}{a numeric}
    \item{\code{hspline1}}{a numeric}
    \item{\code{hspline2}}{a numeric}
    \item{\code{hspline3}}{a numeric}
    \item{\code{fpeaceyears}}{a numeric}
    \item{\code{fspline1}}{a numeric}
    \item{\code{fspline2}}{a numeric}
    \item{\code{fspline3}}{a numeric}
    \item{\code{speaceyears}}{a numeric}
    \item{\code{sspline1}}{a numeric}
    \item{\code{sspline2}}{a numeric}
    \item{\code{sspline3}}{a numeric}
  }
}
\references{
Wimmer, Andreas, Lars-Erik Cederman, and Brian Min. 2009. "Ethnic politics and armed conflict: A configurational analysis of a new global data set." American Sociological Review 74(2): 316-337.}
\examples{
data(Wimmer_et_al_EPR)
head(Wimmer_et_al_EPR)
}
\keyword{datasets}
